Refactor to use a main window
